# front/src/app/api-client/model/family-member-request.ts
/**
 * Agenda familial - API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Création / modification d\'un membre de la famille
 */
export interface FamilyMemberRequest { 
    /**
     * Nom affiché
     */
    nom: string;
    /**
     * Couleur hexadécimale #RRGGBB
     */
    couleur: string;
}